added model folder and refactored database initialization
