--- app/src/main/java/com/example/contactapp/ContactDatabase.kt
@@ -1,9 +1,0 @@
-package com.example.contactapp
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(entities = [Contacts::class], version = 1, exportSchema = false)
-abstract class ContactDatabase: RoomDatabase() {
-    abstract fun contactDao(): ContactsDao
-}
