+++ app/src/main/java/com/example/contactapp/model/ContactDatabase.kt
@@ -1,0 +1,28 @@
+package com.example.contactapp.model
+
+import android.content.Context
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
+
+@Database(entities = [Contacts::class], version = 1, exportSchema = false)
+abstract class ContactDatabase : RoomDatabase() {
+    abstract fun contactDao(): ContactsDao
+
+    companion object {
+        @Volatile
+        private var INSTANCE: ContactDatabase? = null
+
+        fun getDatabase(context: Context): ContactDatabase {
+            return INSTANCE ?: synchronized(this) {
+                val instance = Room.databaseBuilder(
+                    context.applicationContext,
+                    ContactDatabase::class.java,
+                    "contact_database"
+                ).build()
+                INSTANCE = instance
+                instance
+            }
+        }
+    }
+}
