+++ app/src/main/java/com/example/contactapp/model/ContactRepository.kt
@@ -1,4 +1,4 @@
-package com.example.contactapp
+package com.example.contactapp.model
 
 import kotlinx.coroutines.flow.Flow
 
@@ -6,15 +6,15 @@ class ContactRepository(private val contactDao: ContactsDao) {
 
     val allContacts: Flow<List<Contacts>> = contactDao.getAllContacts()
 
-    suspend fun insert(contact:Contacts) {
+    suspend fun insert(contact: Contacts) {
         contactDao.insert(contact)
     }
 
-    suspend fun update(contact:Contacts) {
+    suspend fun update(contact: Contacts) {
         contactDao.update(contact)
     }
 
-    suspend fun delete(contact:Contacts) {
+    suspend fun delete(contact: Contacts) {
         contactDao.delete(contact)
     }
 }
