+++ app/src/main/java/com/example/contactapp/model/Contacts.kt
@@ -1,4 +1,4 @@
-package com.example.contactapp
+package com.example.contactapp.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
